pyproject.toml
[project]
name = "maxflash"
version = "1.0.0"
description = "Integrated Crypto Trading System with Smart Money Concepts"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "scikit-learn>=1.5.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "ccxt>=4.5.0",
    "cachetools>=5.4.0",
    "websocket-client>=1.7.0",
    "pydantic>=2.9.0",
    # ML and Async Trading Dependencies
    "freqtrade>=2024.1",
    "ccxt[async]>=4.5.0",
    "pandas-ta>=0.3.14",
    "tensorflow>=2.18.0",
    "keras>=3.0.0",
    "vectorbt>=0.26.0",
    "aiohttp>=3.11.0",
    "aiodns>=3.5.0",
    "asyncio>=3.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "pytest-timeout>=2.3.0",
    "ruff>=0.8.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]
web = [
    "dash>=2.18.0",
    "dash-bootstrap-components>=1.6.0",
    "dash-table>=5.0.0",
    "plotly>=5.24.0",
    "requests>=2.32.0",
    "gunicorn>=23.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]
visualization = [
    "matplotlib>=3.7.0",
    "plotly>=5.14.0",
]
discord = [
    "discord.py>=2.3.0",
]
ta-lib = [
    "TA-Lib>=0.4.25",
]

[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "N",   # pep8-naming
    "S",   # flake8-bandit (security)
    "T20", # flake8-print
    "PT",  # flake8-pytest-style
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports
"tests/*" = ["ARG001", "S101"]  # test fixtures, use of assert

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "ccxt.*",
    "TA-Lib.*",
    "discord.*",
    "dash.*",
    "plotly.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "smoke: Smoke tests - quick checks for basic functionality",
]

[tool.coverage.run]
source = ["indicators", "utils", "strategies"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

